+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/calculation/CancellationException.java
@@ -1,0 +1,4 @@
+package com.refinedmods.refinedstorage.api.autocrafting.calculation;
+
+class CancellationException extends Exception {
+}