+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/storagemonitor/AutocraftingStorageMonitorContainerMenu.java
@@ -1,12 +1,14 @@
 package com.refinedmods.refinedstorage.common.storagemonitor;
 
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CancellationToken;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.Preview;
-import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewProvider;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.Actor;
+import com.refinedmods.refinedstorage.common.api.autocrafting.CancelablePreviewProvider;
 import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
+import com.refinedmods.refinedstorage.common.autocrafting.PendingAutocraftingRequests;
 import com.refinedmods.refinedstorage.common.autocrafting.preview.AutocraftingPreviewContainerMenu;
 import com.refinedmods.refinedstorage.common.autocrafting.preview.AutocraftingRequest;
 import com.refinedmods.refinedstorage.common.content.Menus;
@@ -19,9 +21,10 @@ import javax.annotation.Nullable;
 import static java.util.Objects.requireNonNull;
 
 public class AutocraftingStorageMonitorContainerMenu extends AutocraftingPreviewContainerMenu
-    implements PreviewProvider {
+    implements CancelablePreviewProvider {
     @Nullable
     private final StorageMonitorBlockEntity storageMonitor;
+    private final PendingAutocraftingRequests pendingAutocraftingRequests = new PendingAutocraftingRequests();
 
     public AutocraftingStorageMonitorContainerMenu(final int syncId, final PlatformResourceKey resource) {
         super(Menus.INSTANCE.getAutocraftingStorageMonitor(), syncId, getRequests(resource));
@@ -42,20 +45,36 @@ public class AutocraftingStorageMonitorContainerMenu extends AutocraftingPreview
     }
 
     @Override
-    public CompletableFuture<Optional<Preview>> getPreview(final ResourceKey resource, final long amount) {
-        return requireNonNull(storageMonitor).getPreview(resource, amount);
+    public CompletableFuture<Optional<Preview>> getPreview(final ResourceKey resource, final long amount,
+                                                           final CancellationToken cancellationToken) {
+        final CompletableFuture<Optional<Preview>> previewRequest =
+            requireNonNull(storageMonitor).getPreview(resource, amount, cancellationToken);
+        pendingAutocraftingRequests.add(previewRequest, cancellationToken);
+        return previewRequest;
     }
 
     @Override
-    public CompletableFuture<Long> getMaxAmount(final ResourceKey resource) {
-        return requireNonNull(storageMonitor).getMaxAmount(resource);
+    public CompletableFuture<Long> getMaxAmount(final ResourceKey resource, final CancellationToken cancellationToken) {
+        final CompletableFuture<Long> maxAmountRequest = requireNonNull(storageMonitor).getMaxAmount(resource,
+            cancellationToken);
+        pendingAutocraftingRequests.add(maxAmountRequest, cancellationToken);
+        return maxAmountRequest;
     }
 
     @Override
     public CompletableFuture<Optional<TaskId>> startTask(final ResourceKey resource,
                                                          final long amount,
                                                          final Actor actor,
-                                                         final boolean notify) {
-        return requireNonNull(storageMonitor).startTask(resource, amount, actor, notify);
+                                                         final boolean notify,
+                                                         final CancellationToken cancellationToken) {
+        final CompletableFuture<Optional<TaskId>> taskRequest = requireNonNull(storageMonitor).startTask(resource,
+            amount, actor, notify, cancellationToken);
+        pendingAutocraftingRequests.add(taskRequest, cancellationToken);
+        return taskRequest;
+    }
+
+    @Override
+    public void cancel() {
+        pendingAutocraftingRequests.cancelAll();
     }
 }