+++ CHANGELOG.md
@@ -7,6 +7,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.0.0-beta.3] - 2025-07-22
+
+### Changed
+
+-   Closing the Crafting Preview window (via cancel or escape) will now properly cancel pending calculations.
+-   Autocrafting requests that take too long will be automatically stopped after 5 seconds.
+-   It is no longer possible to queue an unlimited amount of autocrafting requests when the autocrafting system is busy. 
+
 ## [2.0.0-beta.2] - 2025-04-06
 
 ### Changed
@@ -941,7 +949,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   The Grid can now use smooth scrolling.
 -   The Grid now has syntax highlighting for the search query.
 
-[Unreleased]: https://github.com/refinedmods/refinedstorage2/compare/v2.0.0-beta.2...HEAD
+[Unreleased]: https://github.com/refinedmods/refinedstorage2/compare/v2.0.0-beta.3...HEAD
+
+[2.0.0-beta.3]: https://github.com/refinedmods/refinedstorage2/compare/v2.0.0-beta.2...v2.0.0-beta.3
 
 [2.0.0-beta.2]: https://github.com/refinedmods/refinedstorage2/compare/v2.0.0-beta.1...v2.0.0-beta.2